Refactor video player interface and improve placeholder texture generation
- Changed the return type of draw_window in video_player.hpp from void to bool to indicate success or failure.
- Updated index calculation in create_placeholder_texture in video_player_placebo.cpp for clarity and type safety.
- Adjusted function signature for on_get_app_fullscreen in video_player_placebo.hpp to maintain consistency.
- Modified the type of m_on_set_app_fullscreen in video_player_placebo.hpp from void(bool) to void(int) for better parameter handling.
- Marked external dependencies as dirty to indicate local changes in ColorTextEditorDemo, ImGuiColorTextEdit, and implot.

+++ Project_sdl3_vulkan/code/ui/media/video/player/video_player_placebo.cpp
@@ -181,7 +181,7 @@ bool VideoPlayerPlacebo::create_placeholder_texture() {
     std::vector<uint8_t> pixels(static_cast<size_t>(bytes));
     for (int y = 0; y < h; ++y) {
         for (int x = 0; x < w; ++x) {
-            const size_t idx = static_cast<size_t>((y * w + x) * 4);
+            const size_t idx = (static_cast<size_t>(y) * static_cast<size_t>(w) + static_cast<size_t>(x)) * 4u;
             const bool checker = (((x / 16) + (y / 16)) & 1) != 0;
             const uint8_t r = static_cast<uint8_t>(checker ? 64 : 28);
             const uint8_t g = static_cast<uint8_t>(checker ? 86 : 36);
